Fix password editing and disallow superadmin user deletion

+++ website/util/ClearLogs.php
@@ -1,7 +1,9 @@
 <?php
-session_start();
 
-if($_SESSION['online']){
+session_start();
+require_once("Privilege.php");
+
+if($_SESSION['online'] && Privilege::hasSuperAdmin($_SESSION['privileges'])){
     require_once("../config.php");
 
     $stmt = $db->prepare("TRUNCATE ac_logs");
